Move base_model_context to ert top-level
This is because base_model_context has nothing to do with parsing of
ert.config, and ert.config.parsing is not a public module of ert.config.
