+++ src/ert/base_model_context.py
