+++ src/ert/config/parsing/__init__.py
@@ -1,4 +1,3 @@
-from .base_model_context import BaseModelWithContextSupport
 from .config_dict import ConfigDict
 from .config_errors import ConfigValidationError, ConfigWarning
 from .config_keywords import ConfigKeys
@@ -19,7 +18,6 @@ from .workflow_job_schema import init_workflow_job_schema
 from .workflow_schema import init_workflow_schema
 
 __all__ = [
-    "BaseModelWithContextSupport",
     "ConfigDict",
     "ConfigKeys",
     "ConfigValidationError",
